Refactor settings (GH-36)
